% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierPrerequisites.R
\name{Boxplot_Detect}
\alias{Boxplot_Detect}
\title{Detects outliers using a box-plot}
\usage{
Boxplot_Detect(data_col)
}
\arguments{
\item{data_col}{Any numeric vector that needs to be treated for outliers}
}
\value{
Returns a boolean vector representing TRUE if the data point is an outlier and a FALSE otherwise.
}
\description{
Takes in a data vector and detects the outliers in it using the box-plot statistics
}
