% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierFit.R
\name{ESD_Fit}
\alias{ESD_Fit}
\title{A fit function to detect the outliers using a pre-defined min and max values derived from ESD test}
\usage{
ESD_Fit(data_col, fit_outliers)
}
\arguments{
\item{data_col}{Any numeric vector in which the outliers needs to be detected}

\item{fit_outliers}{A list returned from "OutlierDetectionAndTreatment" that is used to fit the test data.}
}
\value{
Returns a boolean vector representing TRUE if the data point is an outlier and a FALSE otherwise
}
\description{
Detects outliers by taking in a data vector and fit values of the train data
}
