% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierPrerequisites.R
\name{IQRDetect}
\alias{IQRDetect}
\title{Detects outliers using Inter Quartile Range}
\usage{
IQRDetect(data_col)
}
\arguments{
\item{data_col}{Any numeric vector in which the outliers needs to be detected}
}
\value{
Returns a boolean vector representing TRUE if the data point is an outlier and a FALSE otherwise
}
\description{
Takes in a vector and detects outliers using IQR
}
