% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierPrerequisites.R
\name{MAD_Detect}
\alias{MAD_Detect}
\title{Performs Median Absolute Deviation test for outliers on a vector}
\usage{
MAD_Detect(data_col)
}
\arguments{
\item{data_col}{Any numeric vector that needs to be treated for outliers}
}
\value{
Returns a boolean vector representing TRUE if the data point is an outlier and a FALSE otherwise.
}
\description{
Perform Median Absolute Deviation test for up to k potential outliers in a dataset, assuming the data without any outliers come from a normal (Gaussian) distribution.
}
