% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierPrerequisites.R
\name{nSigmaDetect}
\alias{nSigmaDetect}
\title{Detects outliers using mean and standard deviation}
\usage{
nSigmaDetect(data_col)
}
\arguments{
\item{data_col}{Any numeric vector in which the outliers needs to be detected}
}
\value{
Returns a boolean vector representing TRUE if the data point is an outlier and a FALSE otherwise
}
\description{
Takes in a vector and detects outliers using n-Sigma
}
