% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierTreatment.R
\name{OutlierDetectionAndTreatment}
\alias{OutlierDetectionAndTreatment}
\title{Detects outliers in a given data vector}
\usage{
OutlierDetectionAndTreatment(data_col, select)
}
\arguments{
\item{data_col}{Any numeric vector that needs to be treated for outliers}

\item{select}{A metric to mark a data element as an outlier}
}
\value{
Returns an outlier treated data vector using Min-Max capping
}
\description{
Takes in a data vector and detects outliers using multiple outlier-detection techniques:
\itemize{
\item Z-Score Detection
\item Rosner's ESD Test
\item Box Plot Detection
\item Median Absolute Deviation
\item Inter Quantile Range Detection
\item n-Sigma Detection
}
After detecting the outliers using each of these techniques, a data element is detected as outlier only if it detected in more than "select" number of techniques. After detecting the outliers, the data vector is treated for outliers using Min-Max capping.
}
