% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierFit.R
\name{OutlierDetectionAndTreatment_Fit}
\alias{OutlierDetectionAndTreatment_Fit}
\title{A fit function to detect and treat the outliers using the pre-defined statistics for different detection and treatment techniques}
\usage{
OutlierDetectionAndTreatment_Fit(data_col, fit_outliers)
}
\arguments{
\item{data_col}{Any numeric vector in which the outliers needs to be detected}

\item{fit_outliers}{A list returned from "OutlierDetectionAndTreatment" that is used to fit the test data.}
}
\value{
Returns an outlier treated vector
}
\description{
Detects and treats outliers by taking in a data vector and fit values of the train data
}
