% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierFit.R
\name{Outliers_Fit}
\alias{Outliers_Fit}
\title{A fit function to detect and treat the outliers in a dataset using the pre-defined statistics for different detection and treatment techniques}
\usage{
Outliers_Fit(data, fit_outliers, continuous_cols)
}
\arguments{
\item{data}{Any dataset in which the outliers needs to be detected and treated for all continuous fields}

\item{fit_outliers}{A list returned from "bestOutlierTreat" that is used to fit the test data.}

\item{continuous_cols}{A list of continuous columns identified in the train data. Returned from "bestOutlierTreat"}
}
\value{
Returns the original dataset if fit_outliers is NULL and outliers fit dataset otherwise
}
\description{
Detects and treats outliers by taking in a data set and fit values of the train data. Returns the original data if "fit_outliers" is NULL
}
