% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierPrerequisites.R
\name{outliersZ}
\alias{outliersZ}
\title{Performs Z-Score test for outliers on a vector}
\usage{
outliersZ(data_col, zCutOff)
}
\arguments{
\item{data_col}{Any numeric vector that needs to be treated for outliers}

\item{zCutOff}{A z-score threshold exceeding which will be detected as an outlier. Defaulted to 1.96}
}
\value{
Returns a boolean vector representing TRUE if the data point is an outlier and a FALSE otherwise.
}
\description{
Perform Z-Score test in a data vector, assuming the data without any outliers come from a normal (Gaussian) distribution.
}
